requirements.txt
azure-cognitiveservices-nspkg==3.0.1
azure-cognitiveservices-search-nspkg==3.0.1
azure-cognitiveservices-search-websearch==1.0.0
azure-common==1.1.23
azure-nspkg==3.0.2
beautifulsoup4==4.8.1
blis==0.4.1
catalogue==0.0.8
certifi==2019.9.11
chardet==3.0.4
cymem==2.0.3
et-xmlfile==1.0.1
google==2.0.2
idna==2.8
importlib-metadata==1.3.0
isodate==0.6.0
jdcal==1.4.1
lxml==4.4.2
more-itertools==8.0.2
msrest==0.6.10
murmurhash==1.0.2
numpy==1.17.4
oauthlib==3.1.0
openpyxl==3.0.1
plac==1.1.3
preshed==3.0.2
python-docx==0.8.10
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
soupsieve==1.9.4
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
tqdm==4.40.2
urllib3==1.25.6
wasabi==0.4.2
zipp==0.6.0
